Remove duplicated code from stac_fastapi.types (#257)
**Related Issue(s):**
- N/A

**Description:**
We have duplicated code from `stac_fastapi.types` in the code base. This
could introduce hard to find bugs when method signatures in these
abstract classes change and we do not propagate this change (eg. the
path parameter `collection_id` that was introduced recently in the
collection update from the transaction extension). By removing the
duplicated code, we should catch these errors faster and it should be
easier to upgrade to newer `stac_fastapi` versions.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [x] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: jonhealy1 <jonathan.d.healy@gmail.com>