Aggregation (#276)
**Related Issue(s):**

- #245 

**Description:**
Adds support for The [Aggregation
Extension](https://github.com/stac-api-extensions/aggregation) with an
added dependency on the Filter Extension. This enables geo-aggregation
of geometries and points, taking advantage of Opensearch and
Elasticsearch's aggregation engines.

Note that some of the geo-aggregation features will have to be left
untested on the Elasticsearch backend implementation because they
require a commercial license.


TO DO:
- ~~Need to support collection aggregations using the `<collection
ID>/aggregate` route. Add support in the aggregate function in the core
aggregations extension.~~

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [x] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog