refactor aggregate in database_logic (#294)
**Related Issue(s):**
N/A

**Description:**
Refactor `aggregate()` in database logic to allow extending the
supported set of aggregations. The mapping of aggregation name to
Elasticsearch/OpenSearch functionality was in the `aggregate()`
function, which made it difficult to alter the set of supported
aggregations. I moved the mapping to a property of the database logic,
so it can be modified when the database logic is instantiated.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [x] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: Jonathan Healy <jonathan.d.healy@gmail.com>