Adding note on the use of the default `*` use in route authentication dependecies (#325)
**Related Issue(s):**
- #310

**Description:**
Using multiple authentication dependencies on a single route can cause
clashes if not done correctly (merging them). This is particularly
relevant to the default `*` route. My [pull
request](https://github.com/stac-utils/stac-fastapi/pull/766) on
[stac-fastapi](https://github.com/stac-utils/stac-fastapi) to
dynamically merge multiple dependencies was rejected and I don't think
it's possible to replicate it here. So I've added a note to the
dependency documentation to try to explain what to do in this case.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [x] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: Jonathan Healy <jonathan.d.healy@gmail.com>