Add support for optional enum queryables (#390)
**Description:**

This enables support for generating "enum" fields for selected
queryables. There is also now a concept of optional queryable parameters
- so it's possible to indicate enum fields without requiring that field
to be present in all collections. This enum generation is very efficient
as it's basically an index-only scan.

```
    "platform": {
      "description": "Satellite platform identifier",
      "title": "Platform",
      "type": "string",
      "enum": [
        "sentinel-2a"
      ]
    },
```

Other small changes include adding the previously missing
get_items_mapping abstract to the base database logic. I also got
confused by the existence of "OS_HOST" and "OS_PORT" variables in the
Makefile which appear to be unused, so I simply removed them too.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [ ] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: Jonathan Healy <jonathan.d.healy@gmail.com>