Added support for searching large amount of indices (#412)
**Description:**
When searching the catalog with the `/search`-endpoint, a `GET
/<indices>/_search` request is done with all indices listed in the URL
path. However when doing such a search on a large amount of indices, it
is possible that the size of the endpoint exceeds Elasticsearch’s
maximum allowed HTTP line length (4096 bytes), resulting in the
following error:

`{"code":"RequestError","description":"RequestError(400,
'too_long_http_line_exception', 'An HTTP line is larger than 4096
bytes.')"}`

The solution in this commit moves the indices from the endpoint to the
body of the request once the amount of indices passes a certain
threshold. The indices of the endpoint will be replaced by
`ITEM_INDICES`. Since the query still filters on the correct indices,
this change preserves the behavior while avoiding the URL length
limitation.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [ ] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: Stijn Caerts <stijn.caerts@vito.be>