fix: ensure datetime in post request is normalized (#543)
**Description:**

Currently, in `POST` requests to `/search`, with datetime values are not
normalized as they are in `GET` requests to `/search`, which can cause
different results to be returned for the same datetime filter. For
example:
`POST` request to `/search` with
`datetime=2025-11-20T09:54:38Z/2025-11-20T09:54:38Z` returns an item
with `2025-11-20T09:54:38.196614Z`, whereas a GET request with the same
datetime filter does not which is expected behavior. This PR ensures
that datetime values in `POST` requests are normalized in the same way
as in `GET` requests.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [ ] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog

---------

Co-authored-by: Yuri Zmytrakov <yzmytrakovNB@yzmytrakovNB.local>