Add parent ids list to collections, delete collection for collections belonging to catalogs (#554)
**Related Issue(s):**

- #308 
- #520 
-
https://github.com/Healy-Hyperspatial/stac-api-extensions-catalogs-endpoint/issues/3

**Description:**

- Added DELETE `/catalogs/{catalog_id}/collections/{collection_id}`
endpoint to support removing collections from catalogs. When a
collection belongs to multiple catalogs, it removes only the specified
catalog from the collection's parent_ids. When a collection belongs to
only one catalog, the collection is deleted entirely.
- Added `parent_ids` internal field to collections to support
multi-catalog hierarchies. Collections can now belong to multiple
catalogs, with parent catalog IDs stored in this field for efficient
querying and management.

**PR Checklist:**

- [x] Code is formatted and linted (run `pre-commit run --all-files`)
- [x] Tests pass (run `make test`)
- [x] Documentation has been updated to reflect changes, if applicable
- [x] Changes are added to the changelog